+++ ribbon/src/main/java/feign/ribbon/LBClient.java
@@ -33,19 +33,21 @@ import java.net.URI;
 import java.util.Collection;
 import java.util.Map;
 
-class LBClient
+public final class LBClient
     extends AbstractLoadBalancerAwareClient<LBClient.RibbonRequest, LBClient.RibbonResponse> {
 
-  private final Client delegate;
   private final int connectTimeout;
   private final int readTimeout;
   private final IClientConfig clientConfig;
 
-  LBClient(Client delegate, ILoadBalancer lb, IClientConfig clientConfig) {
+  public static LBClient create(ILoadBalancer lb, IClientConfig clientConfig) {
+    return new LBClient(lb, clientConfig);
+  }
+
+  LBClient(ILoadBalancer lb, IClientConfig clientConfig) {
     super(lb, clientConfig);
     this.setRetryHandler(RetryHandler.DEFAULT);
     this.clientConfig = clientConfig;
-    this.delegate = delegate;
     connectTimeout = clientConfig.get(CommonClientConfigKey.ConnectTimeout);
     readTimeout = clientConfig.get(CommonClientConfigKey.ReadTimeout);
   }
@@ -62,7 +64,7 @@ class LBClient
     } else {
       options = new Request.Options(connectTimeout, readTimeout);
     }
-    Response response = delegate.execute(request.toRequest(), options);
+    Response response = request.client().execute(request.toRequest(), options);
     return new RibbonResponse(request.getUri(), response);
   }
 
@@ -82,8 +84,10 @@ class LBClient
   static class RibbonRequest extends ClientRequest implements Cloneable {
 
     private final Request request;
+    private final Client client;
 
-    RibbonRequest(Request request, URI uri) {
+    RibbonRequest(Client client, Request request, URI uri) {
+      this.client = client;
       this.request = request;
       setUri(uri);
     }
@@ -97,8 +101,12 @@ class LBClient
           .request();
     }
 
+    Client client() {
+      return client;
+    }
+
     public Object clone() {
-      return new RibbonRequest(request, getUri());
+      return new RibbonRequest(client, request, getUri());
     }
   }