+++ ribbon/src/main/java/feign/ribbon/RibbonClient.java
@@ -1,11 +1,8 @@
 package feign.ribbon;
 
 import com.netflix.client.ClientException;
-import com.netflix.client.ClientFactory;
 import com.netflix.client.config.CommonClientConfigKey;
 import com.netflix.client.config.DefaultClientConfigImpl;
-import com.netflix.client.config.IClientConfig;
-import com.netflix.loadbalancer.ILoadBalancer;
 import feign.Client;
 import feign.Request;
 import feign.Response;
@@ -13,12 +10,12 @@ import java.io.IOException;
 import java.net.URI;
 
 /**
- * RibbonClient can be used in Fiegn builder to activate smart routing and resiliency capabilities
+ * RibbonClient can be used in Feign builder to activate smart routing and resiliency capabilities
  * provided by Ribbon. Ex.
  *
  * <pre>
- * MyService api = Feign.builder.client(new RibbonClient()).target(MyService.class,
- * "http://myAppProd");
+ * MyService api = Feign.builder.client(RibbonClient.create()).target(MyService.class,
+ *     &quot;http://myAppProd&quot;);
  * </pre>
  *
  * Where {@code myAppProd} is the ribbon client name and {@code myAppProd.ribbon.listOfServers}
@@ -27,13 +24,35 @@ import java.net.URI;
 public class RibbonClient implements Client {
 
   private final Client delegate;
+  private final LBClientFactory lbClientFactory;
 
-  public RibbonClient() {
-    this.delegate = new Client.Default(null, null);
+  public static RibbonClient create() {
+    return builder().build();
   }
 
+  public static Builder builder() {
+    return new Builder();
+  }
+
+  /**
+   * @deprecated Use the {@link RibbonClient#create()}
+   */
+  @Deprecated
+  public RibbonClient() {
+    this(new Client.Default(null, null));
+  }
+
+  /**
+   * @deprecated Use the {@link RibbonClient#create()}
+   */
+  @Deprecated
   public RibbonClient(Client delegate) {
+    this(delegate, new LBClientFactory.Default());
+  }
+
+  RibbonClient(Client delegate, LBClientFactory lbClientFactory) {
     this.delegate = delegate;
+    this.lbClientFactory = lbClientFactory;
   }
 
   @Override
@@ -42,7 +61,8 @@ public class RibbonClient implements Client {
       URI asUri = URI.create(request.url());
       String clientName = asUri.getHost();
       URI uriWithoutHost = URI.create(request.url().replace(asUri.getHost(), ""));
-      LBClient.RibbonRequest ribbonRequest = new LBClient.RibbonRequest(request, uriWithoutHost);
+      LBClient.RibbonRequest ribbonRequest =
+          new LBClient.RibbonRequest(delegate, request, uriWithoutHost);
       return lbClient(clientName)
           .executeWithLoadBalancer(ribbonRequest, new FeignOptionsClientConfig(options))
           .toResponse();
@@ -55,9 +75,7 @@ public class RibbonClient implements Client {
   }
 
   private LBClient lbClient(String clientName) {
-    IClientConfig config = ClientFactory.getNamedConfig(clientName);
-    ILoadBalancer lb = ClientFactory.getNamedLoadBalancer(clientName);
-    return new LBClient(delegate, lb, config);
+    return lbClientFactory.create(clientName);
   }
 
   static class FeignOptionsClientConfig extends DefaultClientConfigImpl {
@@ -72,5 +90,29 @@ public class RibbonClient implements Client {
 
     @Override
     public void loadDefaultValues() {}
+  }
+
+  public static final class Builder {
+
+    Builder() {}
+
+    private Client delegate;
+    private LBClientFactory lbClientFactory;
+
+    public Builder delegate(Client delegate) {
+      this.delegate = delegate;
+      return this;
+    }
+
+    public Builder lbClientFactory(LBClientFactory lbClientFactory) {
+      this.lbClientFactory = lbClientFactory;
+      return this;
+    }
+
+    public RibbonClient build() {
+      return new RibbonClient(
+          delegate != null ? delegate : new Client.Default(null, null),
+          lbClientFactory != null ? lbClientFactory : new LBClientFactory.Default());
+    }
   }
 }