+++ ribbon/src/test/java/feign/ribbon/LBClientFactoryTest.java
@@ -1,0 +1,17 @@
+package feign.ribbon;
+
+import static org.junit.Assert.assertEquals;
+
+import com.netflix.client.ClientFactory;
+import org.junit.Test;
+
+public class LBClientFactoryTest {
+
+  @Test
+  public void testCreateLBClient() {
+    LBClientFactory.Default lbClientFactory = new LBClientFactory.Default();
+    LBClient client = lbClientFactory.create("clientName");
+    assertEquals("clientName", client.getClientName());
+    assertEquals(ClientFactory.getNamedLoadBalancer("clientName"), client.getLoadBalancer());
+  }
+}